Extract the Bug class to a separate class
