+++ src/main/java/sorald/miner/MineSonarWarnings.java
@@ -17,9 +17,9 @@ import java.util.stream.Stream;
 import org.apache.commons.io.FileUtils;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
-import org.sonar.java.AnalyzerMessage;
 import org.sonar.plugins.java.api.JavaFileScanner;
 import sorald.Constants;
+import sorald.sonar.Bug;
 import sorald.sonar.RuleVerifier;
 
 public class MineSonarWarnings {
@@ -203,8 +203,8 @@ public class MineSonarWarnings {
                 }
             }
             for (JavaFileScanner javaFileScanner : SONAR_CHECK_INSTANCES) {
-                Set<AnalyzerMessage> issues = RuleVerifier.analyze(filesToScan, javaFileScanner);
-                warnings.putIfAbsent(javaFileScanner.getClass().getSimpleName(), issues.size());
+                Set<Bug> bugs = RuleVerifier.analyze(filesToScan, javaFileScanner);
+                warnings.putIfAbsent(javaFileScanner.getClass().getSimpleName(), bugs.size());
             }
         } catch (Exception e) {
             e.printStackTrace();
